refactor: 놓친 bookId, clubId를 Int -> Long으로 변경

+++ android/shared/src/commonMain/kotlin/com/yeobaek/data/model/BookDetailModel.kt
@@ -2,7 +2,7 @@ package com.yeobaek.data.model
 
 data class BookDetailModel(
     val authors: List<String>,
-    val bookId: Int,
+    val bookId: Long,
     val chapters: List<ChapterModel>,
     val passageCount: Int,
     val publishedYear: Int,
